Moved demo templates and JS to new file.
