+++ templates/demo_operation_template.html
@@ -1,0 +1,17 @@
+<!DOCTYPE html>
+<html>
+
+    <head>
+        <title>Knitlib operation Template</title>
+
+        <script type="text/javascript" src="/static/knitlib_js_client.js"></script>
+
+    </head>
+
+    <body>
+        <p>Check out the console to test operation.</p>
+
+        <script src="//code.jquery.com/jquery-1.11.3.js"></script>
+    </body>
+
+</html>
